projects/summarizer-agent/requirements.txt
python-dotenv
langchain
openai
duckduckgo-search
